Added a note about EMR and Redshift support
This is a big deal right now. When you use a custom classifier essentially it will not work in EMR or Redshift. In particular not working in EMR means we can't use Hive for ETL. Since Athena doesn't currently support INSERT statements this means that custom classifiers are read only and are a logical dead end for data processing.