Clarify communication direction for AWS Glue VPC Endpoints
The documentation as it stands is misleading. Creation of a VPC Endpoint for Glue does not (as far as I understand it) facilitate connections from Glue jobs to your VPC resources - rather, from the VPC to the AWS Glue API. The case of Glue to VPC communications is handled by Glue Network Connections.